requirements.txt
nonebot
aiohttp
aiocqhttp
requests
lxml
pixivpy